examples/Aduc-app-3d/Aduc-Sdr_Novim/requirements.txt
#accelerate
gradio
google-generativeai
Pillow
ffmpeg-python
sentence_transformers
#transformers
sentencepiece
numpy
torchvision
huggingface_hub>=0.20.0
spaces
opencv-python
imageio
imageio-ffmpeg
einops
timm
safetensors

av
git+https://github.com/huggingface/diffusers.git@main
torch
peft
#diffusers==0.31.0
transformers

accelerate
git+https://github.com/ToTheBeginning/facexlib.git